Refactored install and uninstall plugin tasks. All NPM logic moved to new file called npm.ts. Detect whether a version was specified in the plugin name to install
